make Wallet - Send
Before we used pre instanciated Client with a connector.
This allowed as to use connection pool and connections with keep-alive.
Unfortunately Client is not Send.  So we couldn't share Wallet between
threads. This applies also to Cron, Fsm, FsmPayout since they include
Wallet.

In this PR we create Client only when we need it.